Break coupling between ML2 and L3 during delete operation
This is an initial attempt at breaking out the L3 logic from the ML2
framework as much as possible. This patch takes care of the notification
to the L3 agent(s), after a port has been deleted. Both base L3 and
L3+DVR operations are affected.

Related-blueprint: services-split
Related-blueprint: plugin-interface-perestroika

Change-Id: Ic41b87246ec4d89c2da49afaaccaeb8bd041dcb9