Scope state reporting rpc api using a messaging namespace
This patch does a couple of things.  First it adds docstrings to the
client/server pair of the rpc interface used by an agent to report
state back to the plugin.  The docs tell you where the other side of
the interface is found in the code, and where docs are that give more
info on the rules for changing them.

The second thing done in this patch is to scope this interface using a
messaging namespace.  Right now some plugins expose several interfaces
via the default namespace.  This effectively means they are a single
API and should be managed with a single version stream.  It's much
more managable to just treat these as separate interfaces and this
change makes that explicit and functionally true.  Now when a method
is invoked, the only classes considered for handling that request will
be ones marked with the right namespace.

Part of blueprint rpc-docs-and-namespaces.

Change-Id: I832db5c81780d98157a68d4abaf9787ba00adbeb