Setup br-tun in secure fail mode to avoid broadcast storms
When not creating br-tun in secure fail mode, there are chances to
get a broadcast storm from br-tun.

For example, this occurs when at least three nodes have the br-tun
OpenFlow rules reset in and a broadcast/multicast packet enters br-tun.

This can happen if:
  * openvswitch is restarted, until the agent reloads the Openflow rules.
  * during neutron-openvswitch-agent restart, br-tun is reset, and there
is a few seconds timeframe where tunnel endpoints are plugged and OF
rules are reset.

Secure fail mode doesn't forward traffic by default if no rule is hit.

For stable/juno: create bridge operation is implemented with the secure_mode
parameter, since we don't have ovsdb transactions in juno, and we need
to creation+fail_mode secure to be atomic to avoid race conditions.

Closes-bug: #1421232

Conflicts:
	neutron/agent/linux/ovs_lib.py
	neutron/tests/functional/agent/test_ovs_lib.py

Change-Id: Iba5ded14179156decb16dcd4b898c026660f9653