Monkey patch all the code inside neutron/cmd/eventlet/...
The directory is initially empty.

We are going to maintain entry points for all services and agents that
run in eventlet mode in this directory, and monkey patch them from
there, instead of spreading monkey_patch() calls throughout the library.

This will guarantee us that all the services that are maintained in this
part of the tree monkey patch stdlib properly, before doing any other
imports.

This is also useful to track which parts of the project require
eventlet.  This will later help to migrate services one by one out of
eventlet to real threads in case we decide to move this direction.

Related-Bug: #1417386
Related-Bug: #1418541
Change-Id: I2bc16ca4422c01d64e9fac4910214dbb0d0326ff