VMWare NSXv: id fields should be nullable
To handle concurrency issues when two neutron nodes initialize at the same
time, we use the following solution:
- Each neutron node attempts to create a record with id value=NULL
- One which inserts successfully, resumes to create the resource and once
the resource is created, it updates its id in the database
- One which fails with duplicate entry exception, busy-waits until the id
field is populated, of terminates on timeout.

Relevant resources are internal edges, and internal networks.

To implement the above, id fields should be nullable.

Change-Id: Iff94678ab341ccb1310d6b6efc53a52bd3bb9240
Partially-Implements: blueprint vmware-nsx-v