Make del_fdb_flow() idempotent
Addresses an issue where del_fdb_flow() gets a duplicate call for
the same port, which results in a KeyError. This change makes the
call more idempotent, so that it doesn't cause other follow-on
errors as a result of the uncaught exception.

Change-Id: I7a3a34dd654e143dee06c3e4642e859211a312ca
Closes-bug: #1421105
(cherry picked from commit 841b2f58f375df53b380cf5796bb31c82cd09260)