Update default tox envlist to match voting gate checks
Remove not-yet supported py3* environments.

Note the removed environments (py33,py34) are still supported via
explicit `tox -e`, it's only the default set that is changed here.

ObPoliticalDisclaimer:

This in no way diminishes the importance of py3 support.  In fact, the
intention is quite the opposite:

One day py3x will be supported and the py3x test will be expected to
pass.  Meanwhile, if everyone has been trained to run selective `tox
-epy27` tests then they will be unaware of the py3x status change and
won't be testing their code against py3 - or alternatively they will
believe that py3x failures are "normal" and to be ignored.

The hope with doing this change now is that we will (re)train people
that the default list of tox tests are expected to pass, and any
failures need to be investigated.

Change-Id: I992e1d6b35aa7c516c9ce14e606bbe32e895a4f0