Capture progress chart for vendor decomposition
This patch proposes a way to capture the current status of affairs
for the core/vendor decomposition to help developers and distributors
understand where the Neutron team is in the effort of implementing
the specification.

Initially this was being captured in a Google Doc, but it became
evident that it is best to keep this information as close to the code
as possible. This is very much a collective effort and everyone is
welcome to contribute to keeping this chart current and accurate. In
fact, patches targeting decomposition work will have to address this
table in the same context of the patch that affects the code being
decomposed.

This is not intended to be a documentation tool, or a user reference to
what is and is not a Neutron driver or plugin.

Partial-implements: blueprint core-vendor-decomposition

Change-Id: Ib9a5fa1c06edc837625d55f0ed71d814ee9f7913