Auto-update gateway port after subnet-create
(Patch set #6 for the multiple-ipv6-prefixes blueprint)

In the multi-prefix scenario, one can add two subnets
to an external gateway port by adding the two subnets
to the external network and using router-gateway-set.

However, if there is only one subnet on the port and
the user wishes to add another later, it is desirable
to have the newly-created external subnet automatically
added to the port. This patch adds this functionality.

Change-Id: I9395834f673038dc23b25eaeefe14895fe154e0e
Partially-implements: blueprint multiple-ipv6-prefixes