Make DHCP tests cleanup neutron manager reference
The DHCP scheduler unit tests occasionally call methods
on the core plugin via the neutron manager (e.g. in report_state).
However, they weren't calling the normal unit test cleanup routines
for tests that call the core plugin so they were leaving stale
references in the manager.

This patch uses the setup_coreplugin helper that sets up the
configuration for the core plugin and adds the cleanup call
to ensure that the neutron manager is cleaned up at the end
of each test.

Closes-Bug: #1434278
Change-Id: Ic35629659574121c739574c1ed39364d6977c3ea