Pass correct port ID back to RPC caller
The previous response to get_device_details calls was returning
whatever the caller requested as the port_id in the response.
This was only correct in the case where the port_id was used
directly. In cases where device names were passed in, there was
no way to retrieve the full port ID.

This corrects that behavior by using the port ID from the database
and adds tests to ensure the behavior remains correct.

Closes-Bug: #1443714
Change-Id: Ibfc7b6659a29e892dfe6e83bd9340feb40e920dd