Handle SIGHUP in dhcp and l3 agents
All launchers implemented in common.service require each service to
implement reset method because it is called in case a process
receives a SIGHUP.

This change adds the reset method to neutron.service.Service class
which is used to start dhcp and l3 agents.

Now dhcp and l3 agents don't die on receiving SIGHUP and support
reloading policy_path and logging options in config.

Partial-Bug: #1276694
Change-Id: I96010e44928a665bea546865b2c81bde4ed0adf2