Enable ARP spoofing prevention by default
Turn on the ARP spoofing prevention added in
I7c079b779245a0af6bc793564fa8a560e4226afe by
default. It was disabled by default since it
was going into Kilo at the last minute and we
didn't want to risk shipping with a default
that might have broken an edge case that we
didn't consider.

This patch enables it by default since there
shouldn't be any need to have it disabled.

Change-Id: Id17939914ebf8292dce76ccb7d0f6486c91f49e5