lb-agent: ensure tap mtu is the same as physical device
On compute-nodes, Nova creates the bridge with the tap before
the physical is in the bridge. This causes the tap to have the
default 1500 MTU which may be different to what is on the physical.
With this patch the linuxbridge agent ensures that the MTU on the
tap device is the same as what is on the physical device.

Change-Id: Id1a4f662ec33ca0333c15eb210366bc850d0d54c
Closes-Bug: 1443607