Quota model: use HasTenantId mixin
This change simply changes the Quota model class to obtain
the tenant_id from the mixin class. As the attribute in the
mixin is identical to that in the model there is no need for
a migration.

This patch also removes a reference to quota classes in the
docstring, as Neutron does not implement those. It is good
to be careful when copying and paste code.

Change-Id: Idab15d5ef2ddd2b830a7dcde46990506064535f7
Closes-Bug: #1445169