SystemExit is ok for child processes
DietTestCase catches SystemExit while running tests, interprets it as a
test failure, and then carry on with the next test (without exiting).
This greatly upsets forked child python processes, which may call exit()
legitimately, and expect that to result in process exit.

This change re-raises the SystemExit if the current process ID is not
the original pid.

Change-Id: Ia39a350b562b2856b5588cd73826afb3d072554f