Add notes about official sub-projects.
There was recently a thread on openstack-dev titled "A big tent home
for Neutron backend code."

The thread began here:

    http://lists.openstack.org/pipermail/openstack-dev/2015-April/062310.html

and has roughly ended up here:

    http://lists.openstack.org/pipermail/openstack-dev/2015-April/062853.html

This patch is an attempt to reflect the end of that thread with
updates to docs.

Any further discussion should just continue on openstack-dev to avoid
forking the discussion between openstack-dev and gerrit.

Change-Id: I48dbe8ac69e60fbfd5e5082844004aaf9fdce539