Fix floatingip status for an HA router
When we associate a floatingip in an HA router setup, it is properly
associated. However, when we check the status of the Floating ip, it
is shown as empty.  The L3 agent ha_router implementation never sets
the status of the floating IP when it's added.  This behavior blocks
tempest scenario tests that use FIPs when run against a deployment
with L3 agent HA routers.  This fix sets the FIP status to active
when the FIP is added by an HaRouter instance.

Co-Authored-By: Assaf Muller <amuller@redhat.com>
Closes-Bug: #1449049
Change-Id: I7a3de36b64132e483a927ce9fed30777e84df96a