Switch to dictionary for iptables find
The code to find the matching entry was scanning through a
list of all rules for every rule. This became extremely slow
as the number of rules became large, leading to long delays
waiting for firewall rules to be applied.

This patch switches to the use of a dictionary so the cost
becomes a hash lookup instead of a list scan.

Closes-Bug: #1453264
Closes-Bug: #1455675
(cherry picked from commit 7a3934d982ef29d8851450b5586319201baa0122)

Conflicts:

	neutron/agent/linux/iptables_manager.py

Change-Id: I1e6fe5e50b9c13066c966c252cadc8ed1d08f686