Add a non-mixin function for model queries
This patch simply adds a version of model_query in
neutron.db.common_db_mixin which can be invoked without
having to declare a class which inherits the mixin.

To this aim, model_query_scope has been refactored as well.
As the model query function being introduced in this patch
cannot use model query hooks (and does not need to), the
method was re-implemented rather than bringing out of the
mixin as it has been done for model_query_scope.

This change will allow for developing DB APIs without
having to use the baseDB/mixin classes models used so far.

Related-Blueprint: better-quotas

Change-Id: I7a79980f626e9eaf2775711c8a25f508067e5716