Remove middleware oslo-incubator module
The module was used during Kilo cycle to provide backwards compatibility
for users that upgrade to the release without updating their
api-paste.ini. We have issued the deprecation warning for a cycle now,
so we should be ok to just drop the compatibility layer.

Note that the change may require a notion in release notes to make sure
everyone is notified, even if they don't look through their logs.

DocImpact

Change-Id: I41693f4613b5a69a01a33e54f90e82177f42e1af