NSX QoS ext: RXTX factor can be decimal
In Nova flavors it is ok to specify a decimal RXTX factor.
For this reason when applying QoS to a port Neutron should not
convert this factor to an integer value, but simply ensure
it's a valid float number and positive.

This patch ialso drops the conversion to integer that happens
in the NSX-mh plugin but also ensures that the resulting maximum
bandwidth after applying the RXTX factor is still an integer value.

Closes-Bug: #1463363
Patch applied from commit: 9f61b082093fde7db135f400d7d99724c6c01a43
Change-Id: I9f5e166e349de0703a8711fa73a127e90059f506