NSX QoS ext: RXTX factor can be decimal
In Nova flavors it is ok to specify a decimal RXTX factor.
For this reason when applying QoS to a port Neutron should not
convert this factor to an integer value, but simply ensure
it's a valid float number and positive.

Cherry-picked from commit: 538a7bf3c73d968185590d81a9b3a6523190aeb4
Partial-Bug: #1463363

Change-Id: I983123ef7fd8f1b52b358aff3b579459fce63033