Fix l3 agent to not create already deleted router
In case router is deleted during l3 agent resync,
the "deleted" event is processed with higher priority, then
resync event for the router may be processed which will recreate
already deleted router.
This happens due to timestamp not being properly updated for deleted
router in router processor.
The fix adds timestamp update for deleted router.

Functional test will be updated in a follow-up patch

Logging was improved to make debugging a bit easier.

Closes-Bug: #1455439
Change-Id: I2d060064acccc10591a3d90be9011f116548cfce