Delete FIP agent gateway port with external gw port
FIP agent gateway ports are associated with external
networks and specific host.

Today FIP agent gateway ports are deleted for
every floatingip associate and disassociate. This
introduces race conditions in the port delete and also
un-necessary access to the db.

This patch will delete the FIP agent gateway port when
the last gateway port of the external network is deleted.

The child patch linked to this parent patch will clean
up the FIP agent gateway port delete when associate,
disassociate and delete of floatingip happens.

This should also cover the case when an agent for some
reason was unable to request agent gw port delete.
(agent died).

Related-Bug: #1408855
Related-Bug: #1468007
Related-Bug: #1450982

Change-Id: I6637a771e6a6ce74e848cb74b779043e16a54a84