Cleanup the fip agent gateway port delete routines
Based on the parent patch, right now the Floatingip
agent gateway ports will only be deleted when the
last gateway port associated with the external
network is deleted.

The Floatingip agent gateway port will not be deleted
for every floatingip dis-association and deletion.

The Floatingip agent gateway port was created on all
nodes as a substitute for the gateway port. So it
makes sense to delete those ports only when the last
gateway port on the external network is deleted.

The agent should be able to delete the floatingip agent
gateway port on a given external network when it is not
required.

This would substantially reduce the burden on the server
to validate, read and delete the port form the DB.

Change-Id: Ie561b19a2e58a2a563d79b75421e9e24c70f36f9
Closes-Bug: #1468007
Closes-Bug: #1408855
Closes-Bug: #1450982