Don't process network_delete events on OVS agent
There is no reason to reclaim the local VLAN on a network_delete
event since it will be reclaimed when the last port is deleted anyway.

This method was racey with recent port creation events and lead to
occasional traces when it would yank the local VLAN out from underneath
an ongoing port wiring process.

Change-Id: I5eee6175c053006b35b6efe274d27931a9d5d89f
Closes-Bug: #1468251