QoS: db models and migration rules
This patch includes db models and migration rules for initial QoS objects.

Comparing to the spec, it adds two more service tables to maintain links
between networks and ports and their respective policies. We maintain
uniqueness as a unique constraint.

In some parallel world, we could have an additional field for networks
and ports that could be nullable to point to a policy. That said, it
breaks qos isolation a bit, and will also be a bit more painful if and
when we decide to spin out qos service pieces outside the tree.

blueprint quantum-qos-api

Co-Authored-By: Ramanjaneya <ramanjieee@gmail.com>
Co-Authored-By: vikram.choudhary <vikram.choudhary@huawei.com>
Co-Authored-By: Ihar Hrachyshka <ihrachys@redhat.com>
Co-Authored-By: Miguel Angel Ajo <mangelajo@redhat.com>

Change-Id: I55a7dac602e2e770c21b6c7957430cb7115e5bdc