Use last address in v6 allocation pool generation
This reverts commit 1d61c5f736b9df7253a1939fe55f1875f988d1f4,
which reverted ee51ef72d37a02005a7733b7f2faf7236db850a1 due
to an issue with the allocation pools being mis-aligned with
the last address of a v6 subnet.

IPv6 does not have a subnet broadcast address so the last IP
address can be used as a regular address. This patch tweaks the
automatic allocation pool generation to include that last address
on v6 networks.

This was found because
NetworksTestDHCPv6.test_dhcp_stateful_fixedips_outrange makes the
assumption that the default generated pools cover the entire usable
address space.

Related-Bug: #1466322
Closes-Bug: #1468163
Change-Id: I5a2518c819757cd1fb43244901fe3e2f6c576b34