Add extension callbacks support for networks
Add callbacks for extention to the network resource so that interested
extensions can do custom logic when a network is created or updated.

Currently it will be done the same way port notifications are done -
i.e. only in ML2 plugin. We can revisit this in a floow up patch if the
whole notification logic should be moved somewhere else as this merits
further discussion which is out of scope for a mere patch.

This will be utilized in a follow up commit.

Partially-implements: blueprint quantum-qos-api
Change-Id: I38528863e1145caf05fe3b2425511d1c5b5c0f93