Update dhcp host portbinding on failover
When a dhcp server is moved to new host on failover,
tunnel is not created.

When a dhcp server is moved to new host,
1) network-hostagent binding is updated(properly i.e with new host).
2) dhcp port-hostagent binding is not updated
   ( dhcp port is still bound to old host)
If dhcp port-bound agent is different from the new dhcp agent
(which is now taking care of this dhcp port), neutron plugin won't
 notify the l2pop, and hence tunnel is not created.

As after failover, the new agent is taking care of this dhcp port,
update portbinding with the new host. This will allow neutron plugin
to notify l2pop(which will create tunnel).

Change-Id: Ib7d7dcddee005395af116ccd31a43853332ae317
Closes-bug: #1411163
(cherry picked from commit 571453c614577ffd47e23de9a35863d2f2119583)