Allow passing table argument to construct IpRouteCommand
The various methods of IpRouteCommand also work with a non-default
table.  Instead of passing an explicit table argument to each command,
this change allows constructing an instance of IpRouteCommand which
always works on the given table much like we pass the namespace on
creation instead of passing it to each command individually.

Change-Id: I503aa986208347ccc1cb5f37bdb36f77553eee70
Partially-Implements: blueprint address-scopes