Arista ML2 driver should ignore non-vlan networks
Arista ML2 Mech driver for VLANs presently does
not filter out non-vlan type networks (e.g vxlan).
This fix will simply ignore the request if a
non-vlan based network request is seen.

This patch is a juno port for https://review.openstack.org/#/c/199380/8
The cherry-picking had to be done manually because of the decomposition
of the repository

Conflicts:
	neutron/plugins/ml2/drivers/arista/mechanism_arista.py
	neutron/tests/unit/plugins/ml2/drivers/arista/test_mechanism_arista.py

Change-Id: I99ec5c5772a9f1f63cf871a98be50bdd2ba5db62
Closes-Bug: 1472458