Hyper-V: Fix for ICMP rule deletion
With the introduction stateful security rules a change to set localport
to empty for icmp is made during rule cration.
However during deletion there should be a check to match the empty
localport field in case of icmp , since _filter_security_acls method
will not match the rule on hyperv network adapter extended acls which
results in icmp rule not getting deleted and traffic still flows.
With the current fix this issue is addressed.

This fix is targeted in specific to  Juno branch and is different
from kilo branch because of the below reasons:

Hyper-V Security Groups have been heavily refactored in the kilo release
in order to simplify the rule creation process and reduce the complexity

The packaging has also changed and hence it is not possible to backport
the changes directly and needs a fix specific to Juno

Review for Kilo branch : https://review.openstack.org/199978
Change-Id is maintained same as kilo fix to have consistency

Change-Id:  If10d1408bf0f3ed487e394edd8e0b0ea52070a48
Closes-Bug: #1470443