Add versioned object serialize/deserialize for resources RPC
Also switched RPC callback API to consistently receive resource_type
string and not a resource class. This is because for get_info(), we
cannot propagate a class thru RPC but only a string that uniquely
identifies the class. So it would be not optimal to require the server
to discover the corresponding class from the type name passed from the
agent.

Also updated some comments in api/rpc/callbacks directory to reflect
that we handle NeutronObjects, not dicts.

Finally, killed the rule resource registration from QoS plugin and the
rule type from supported resources since it's YAGNI at least now.

Partially-Implements: blueprint quantum-qos-api
Change-Id: I5929338953a2ad7fa68312d79394a306eb0164a2