In Arista ML2 driver Reconfigure VLAN on VM migration
Whenever a VM moves from one compute node to the other, the VLAN on the old
switch interface was not removed and the VLAN was not being provisioned on the
new switch interface. With this patch, the VLANs are provisioned correctly.

This patch is a juno port of https://review.openstack.org/#/c/198182/ with
some changes to account for the driver decomposition.

Change-Id: I658ef87ffd82119f41beda9ba019e29a62dc96ff
Closes-Bug: #1471050