Adds garp_master_repeat and garp_master_refresh to keepalived.conf
This patch aims to solve an issue when, in some cases, the L3 agent might
start before the OVS agent during a node boot process.
At such scenario, the L3 agent will configure the routers before the OVS agent
done syncing with the server and starts processing ovsdb monitor updates.

The result:
1. L3 agent finishes configuring an HA router
2. Keepalived spawned
3. In case of transition to master state, it (keepalived) sends gratuitous ARPs
                                                            ^---The issue
4. The OVS agent has just finished plugging its ports

Meaning, the gratuitous ARP will be lost, but the router acting as master.
Closes-Bug: #1453855

Change-Id: I2cffa7c0ebad20bd50396265aef9fcfa64018744